更新使用token来获取user_id

+++ app/views.py
@@ -673,7 +673,7 @@ class MakeLeave(View):
         user_id = User.objects.get(phone_number=jwt_token['username']).user_id
         user = User.objects.get(user_id=user_id)
         if user.type == 'doctor':
-            doctor_id = Doctor.objects.filter(phone_number=jwt_token['username']).first()
+            doctor_id = Doctor.objects.filter(phone_number=jwt_token['username']).first().doctor_id
             json_str = request.body
             json_obj = json.loads(json_str)
             start_time = json_obj['start_time']
@@ -711,15 +711,24 @@ class MakeLeave(View):
 class CancelLeave(View):
     @method_decorator(logging_check)
     def post(self, request):
-        json_str = request.body
-        json_obj = json.loads(json_str)
-        doctor_id = json_obj['doctor_id']
-        start_time = json_obj['start_time']
-        end_time = json_obj['end_time']
-        leave = Leave.objects.filter(doctor_id_id=doctor_id, start_time=start_time, end_time=end_time).first()
-        if leave:
-            leave.delete()
-        return JsonResponse({"result": "1", "message": "取消预约成功！"})
+        token = request.META.get('HTTP_AUTHORIZATION')
+        jwt_token = jwt.decode(token, settings.JWT_TOKEN_KEY, algorithms='HS256')
+        user_id = User.objects.get(phone_number=jwt_token['username']).user_id
+        user = User.objects.get(user_id=user_id)
+        if user.type == 'doctor':
+            json_str = request.body
+            json_obj = json.loads(json_str)
+            doctor_id = Doctor.objects.filter(phone_number=jwt_token['username']).first().doctor_id
+            start_time = json_obj['start_time']
+            end_time = json_obj['end_time']
+            leave = Leave.objects.filter(doctor_id_id=doctor_id, start_time=start_time, end_time=end_time).first()
+            if leave:
+                leave.delete()
+                return JsonResponse({"result": "1", "message": "取消预约成功！"})
+            else:
+                return JsonResponse({"result": "0", "message": "出错啦！"})
+        else:
+            return JsonResponse({"result": "0", "message": "用户未有权限"})
 
 
 class PatientAppointment(View):
@@ -850,7 +859,6 @@ class DeletePatient(View):
         json_str = request.body
         json_obj = json.loads(json_str)
         patient_id_list = json_obj['patient_ids']
-
         try:
             with transaction.atomic():
                 for patient_id in patient_id_list:
